Move config-related things to the config directory
