+++ packages/react-native-reanimated/src/config/index.ts
@@ -1,0 +1,18 @@
+'use strict';
+
+export type { LoggerConfig, LoggerConfigT } from './ConfigHelper';
+export {
+  addWhitelistedNativeProps,
+  addWhitelistedUIProps,
+  configureReanimatedLogger,
+} from './ConfigHelper';
+export type { FeatureFlag } from './featureFlags';
+export {
+  getDynamicFeatureFlag,
+  getStaticFeatureFlag,
+  setDynamicFeatureFlag,
+} from './featureFlags';
+export {
+  isReducedMotionEnabledInSystem,
+  ReducedMotionManager,
+} from './ReducedMotion';
