+++ packages/react-native-reanimated/src/createAnimatedComponent/AnimatedComponent.tsx
@@ -9,8 +9,8 @@ import { IS_JEST, IS_WEB, logger } from '../common';
 import type { StyleProps } from '../commonTypes';
 import { LayoutAnimationType } from '../commonTypes';
 import { SkipEnteringContext } from '../component/LayoutAnimationConfig';
+import { getStaticFeatureFlag } from '../config';
 import ReanimatedAnimatedComponent from '../css/component/AnimatedComponent';
-import { getStaticFeatureFlag } from '../featureFlags';
 import type { AnimatedStyleHandle } from '../hook/commonTypes';
 import { type BaseAnimationBuilder } from '../layoutReanimation';
 import { SharedTransition } from '../layoutReanimation/SharedTransition';
