+++ packages/react-native-reanimated/src/mutables.ts
@@ -11,7 +11,7 @@ import {
 
 import { IS_JEST, logger, ReanimatedError, SHOULD_BE_USE_WEB } from './common';
 import type { Mutable } from './commonTypes';
-import { getStaticFeatureFlag } from './featureFlags';
+import { getStaticFeatureFlag } from './config';
 import { isFirstReactRender, isReactRendering } from './reactUtils';
 import { valueSetter } from './valueSetter';
